Move highscoreManager to highscore.rs - JIRA-47
